+++ auth/apps/frontend/src/app/(authenticated)/home/page.tsx
@@ -2,7 +2,7 @@
 
 /* * */
 
-import { HomePage } from '@/components/home/HomePage2';
+import { HomePage } from '@/components/home/HomePage';
 
 /* * */
